提交訊息：refactor: 重新命名 'ezban' package 為 'com;
這次的提交將所有的 'ezban' 包名改成了 'com'。
