+++ src/main/java/com/ServletInitializer.java
@@ -1,4 +1,4 @@
-package ezban;
+package com;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
