+++ src/main/java/com/controller/BackstageController.java
@@ -1,4 +1,4 @@
-package ezban.controller;
+package com.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
